Corrections : Add App.scss; DataTable Mapping; common Utils/API.js
