+++ package.json
@@ -8,11 +8,12 @@
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
     "check-prop-types": "^1.1.2",
+    "lodash": "^4.17.21",
     "lucid-ui": "^5.0.0",
     "moxios": "^0.4.0",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-redux": "^7.2.2",
     "react-scripts": "4.0.3",
     "redux": "^4.0.5",
