+++ src/App.scss
@@ -1,0 +1,20 @@
+.App {
+    text-align: center;
+
+    .paginator-container {
+        margin: 25px;
+
+    }
+
+    .App-paginator {
+        display: 'flex';
+        flex-direction: 'row';
+        align-items: 'center';
+        margin-bottom: 10;
+    }
+
+    .lucid-Paginator {
+        justify-content: center;
+        margin: 10px;
+    }
+}
