+++ src/actions/actionTypes.js
@@ -1,0 +1,8 @@
+export const actionTypes = {
+    SET_TODOS: "SET_TODOS",
+    SET_PAGE: "SET_PAGE",
+    SET_PAGE_SIZE: "SET_PAGE_SIZE",
+    SET_TODOS_LIST: "SET_TODOS_LIST",
+    SET_ERROR: "SET_ERROR",
+  };
+
