+++ src/reducers/reducer.js
@@ -1,4 +1,4 @@
-import { actionTypes } from "../actions";
+import { actionTypes } from "../actions/actionTypes";
 
 const initialStore = {
   todos: [],
