+++ src/utils/API.js
@@ -1,0 +1,9 @@
+import axios from "axios";
+export default {
+  fetchTodos: async () => {
+    return await axios
+      .get("https://jsonplaceholder.typicode.com/todos")
+      .then((res) => res)
+      .catch((err) => err);
+  },
+};
